+++ src/test/java/lotto/domain/LottoTest.java
@@ -1,13 +1,11 @@
-package lotto;
+package lotto.domain;
 
-import lotto.domain.Lotto;
-import lotto.domain.LottoResult;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
 
-import static lotto.ValidatorTest.WINNING_NUMBERS;
+import static lotto.exception.ValidatorTest.WINNING_NUMBERS;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
