+++ src/test/java/lotto/service/LottoServiceTest.java
@@ -1,12 +1,11 @@
-package lotto;
+package lotto.service;
 
-import static lotto.ValidatorTest.WINNING_NUMBERS;
+import static lotto.exception.ValidatorTest.WINNING_NUMBERS;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
 import lotto.domain.Lotto;
 import lotto.domain.LottoResult;
-import lotto.service.LottoService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
